requirements.txt
click==7.1.2
Pillow==8.1.2
pyinstaller==4.2
pyserial==3.5
pywin32==300
